Simplified auto parallelism control, added new methods, removed reschedule functions
