+++ src/meta/node/src/server.rs
@@ -676,9 +676,7 @@ pub async fn start_service_as_election_leader(
         sub_tasks.push(GlobalBarrierManager::start(barrier_manager));
 
         if env.opts.enable_automatic_parallelism_control {
-            sub_tasks.push(GlobalStreamManager::start_auto_parallelism_monitor(
-                stream_manager,
-            ));
+            sub_tasks.push(stream_manager.start_auto_parallelism_monitor());
         }
     }
     let (idle_send, idle_recv) = tokio::sync::oneshot::channel();
